build: keep ROM cast rewrite in CMake and scope warnings-as-errors to MSVC
Generate the explicit unsigned casts for the Shruthi ROM tables directly in
ShruthiRomResources.cmake instead of shelling out to a Python helper, so the
build no longer requires a Python interpreter. The generated translation unit
is byte-for-byte identical to the previous output.

Store the envelope trigger flags as -1 rather than a narrowing conversion of
255: the stored bit pattern and the value read back through
modulation_destination() are unchanged, without relying on an
implementation-defined signed conversion. The MSVC pragma that muted 4100,
4244, 4245 and 4838 around the part.h include is no longer needed.

Restrict warnings-as-errors to MSVC, where the tree is warning-free. GCC and
Clang still report pre-existing diagnostics from the untouched upstream
firmware headers, so -Werror there would break the Linux and macOS builds.

Make the remaining conversions in the test suites explicit so the whole tree
builds without warnings.

+++ Tests/Correctness/CorrectnessTestsMain.cpp
@@ -1081,7 +1081,7 @@ void testFmRangeSemantics()
         expect(voice.debug_oscillator_secondary_parameter(0) == range + 24
                    && voice.debug_oscillator_secondary_parameter(1) == range + 24,
                "FM range reaches the native secondary ratio parameter");
-        for (int oscillator = 0; oscillator < 2; ++oscillator)
+        for (uint8_t oscillator = 0; oscillator < 2; ++oscillator)
         {
             const auto increment = voice.debug_oscillator_increment(oscillator);
             if (range == -24)
